requirements.txt
# Core agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM providers
langchain-google-genai>=2.0.0   # Gemini (default)
langchain-openai>=0.2.0          # OpenAI (optional)
langchain-anthropic>=0.3.0       # Claude (optional)

# Embeddings & vector store
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Sparse retrieval
rank-bm25>=0.2.2

# SEC filings ingestion
edgartools>=2.20.0

# Observability
langsmith>=0.1.0

# Evaluation
ragas>=0.2.0
deepeval>=1.4.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
